Fixed issue with Cinemachine cameras
Static Camera, Gameplay Camera, and Target Group all need to exist in the scene you want to play the game in. You'll also need to set the Follow and Look At transforms in the Gameplay Camera to the Target Group transform by dragging it in there in the Inspector. Same for the Cinemachine object references in the Game Manager.